backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
pillow
